+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rewrite the checksum algorithm for readability and simplicity.
+
 ## [1.1.1] - 2023-12-23
 
 ### Fixed
@@ -14,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python bindings:
   - Fix `detectCIC` and `detect_cic_raw` functions not accepting `bytearray`
     objects.
-- Fix some typos
+- Fix some typos.
 
 ## [1.1.0] - 2023-12-22